-Change the hashfunc struct for the OString class to return an unsigned int instead of size_t. Now it matches the rest of the built-in hash functions for OHashTable.
SVN r5226 (trunk)

+++ common/m_ostring.h
@@ -661,7 +661,7 @@ private:
 // ----------------------------------------------------------------------------
 
 template <> struct hashfunc<OString>
-{   size_t operator()(const OString& str) const { return str.mId; } };
+{   unsigned int operator()(const OString& str) const { return static_cast<unsigned int>(str.mId); } };
